+++ changelog.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.0.2] - 2020-12-23
+ 
+### Added
+
+- Supports `pccomponentes` website as new vendor.
+- Adds selenium and chromedriver as requirements to support web JS rendering.
+
 ## [0.0.1] - 2020-12-21
  
 ### Added